Overall performance improvements
Using the ::after technique instead of backdrop-filter is demanding and so is the backdrop filter. I was using them in places were it was not needed. Made a specific CSS class to replace bg-glass-rounded where it is not needed to do a backdrop-filter. That seemed to help the webpage performance overall.

Signed-off-by: Moe Ayoub <mma.ayoub@outlook.com>